Split unsupported target tests
