+++ tests/integration/cli_build/unsupported_targets.rs
@@ -1,155 +1,27 @@
 use std::process::Command;
 
-#[test]
-fn build_command_build_zen_rejects_unsupported_package_targets() {
-    assert_build_zen_command_rejects_unsupported_target_kind(&["build", "build.zen"], "Package");
-}
+#[path = "unsupported_targets/fields.rs"]
+mod fields;
+#[path = "unsupported_targets/kinds.rs"]
+mod kinds;
 
-#[test]
-fn build_command_build_zen_rejects_unsupported_link_targets() {
-    assert_build_zen_command_rejects_unsupported_target_kind(&["build", "build.zen"], "Link");
-}
-
-#[test]
-fn direct_file_command_build_zen_rejects_unsupported_package_targets() {
-    assert_build_zen_command_rejects_unsupported_target_kind(&["build.zen"], "Package");
-}
-
-#[test]
-fn direct_file_command_build_zen_rejects_unsupported_link_targets() {
-    assert_build_zen_command_rejects_unsupported_target_kind(&["build.zen"], "Link");
-}
-
-#[test]
-fn check_command_build_zen_rejects_unsupported_package_targets() {
-    assert_build_zen_command_rejects_unsupported_target_kind(&["check", "build.zen"], "Package");
-}
-
-#[test]
-fn check_command_build_zen_rejects_unsupported_link_targets() {
-    assert_build_zen_command_rejects_unsupported_target_kind(&["check", "build.zen"], "Link");
-}
-
-#[test]
-fn test_command_build_zen_rejects_unsupported_package_targets() {
-    assert_build_zen_command_rejects_unsupported_target_kind(&["test", "build.zen"], "Package");
-}
-
-#[test]
-fn test_command_build_zen_rejects_unsupported_link_targets() {
-    assert_build_zen_command_rejects_unsupported_target_kind(&["test", "build.zen"], "Link");
-}
-
-#[test]
-fn emit_command_build_zen_rejects_unsupported_package_targets() {
-    assert_build_zen_command_rejects_unsupported_target_kind(&["emit", "build.zen"], "Package");
-}
-
-#[test]
-fn emit_command_build_zen_rejects_unsupported_link_targets() {
-    assert_build_zen_command_rejects_unsupported_target_kind(&["emit", "build.zen"], "Link");
-}
-
-#[test]
-fn build_graph_command_rejects_unsupported_package_targets() {
-    assert_build_zen_command_rejects_unsupported_target_kind(
-        &["build-graph", "build.zen"],
-        "Package",
-    );
-}
-
-#[test]
-fn build_graph_command_rejects_unsupported_link_targets() {
-    assert_build_zen_command_rejects_unsupported_target_kind(&["build-graph", "build.zen"], "Link");
-}
-
-#[test]
-fn build_zen_commands_reject_package_fields() {
-    assert_build_zen_commands_reject_gated_target_field("packages", r#"["std"]"#);
-}
-
-#[test]
-fn build_zen_commands_reject_link_fields() {
-    assert_build_zen_commands_reject_gated_target_field("link", r#"["m"]"#);
-}
-
-#[test]
-fn build_zen_commands_reject_unknown_target_fields() {
-    assert_build_zen_commands_reject_unknown_target_field("output_dir", r#""build/app/""#);
-}
-
-fn assert_build_zen_command_rejects_unsupported_target_kind(args: &[&str], target_kind: &str) {
-    let tmp = tempfile::tempdir().expect("create temp dir");
-    std::fs::write(
-        tmp.path().join("build.zen"),
-        format!(
-            r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {{
-    b.add({target_kind} {{ name: "core", root: "src/lib.zen" }})
-    .Ok(b.config())
-}}
-"#
-        ),
-    )
-    .expect("write build.zen");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_zen"))
-        .args(args)
-        .current_dir(tmp.path())
-        .output()
-        .expect("run zen build.zen command");
-
-    assert!(
-        !output.status.success(),
-        "zen {args:?} unexpectedly succeeded: stdout={}, stderr={}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-    assert!(
-        String::from_utf8_lossy(&output.stderr).contains(&format!(
-            "unsupported build target kind `{target_kind}`; supported target kinds are `Executable`, `Test`, and `Library`"
-        )),
-        "expected unsupported target diagnostic, stderr={}",
-        String::from_utf8_lossy(&output.stderr)
-    );
-    assert!(
-        !tmp.path().join("build").exists(),
-        "zen {args:?} should reject unsupported target kinds before creating build outputs"
-    );
-}
-
-fn assert_build_zen_commands_reject_gated_target_field(field: &str, value: &str) {
-    for args in [
+fn all_build_zen_command_args() -> [&'static [&'static str]; 6] {
+    [
         &["build", "build.zen"][..],
         &["build.zen"][..],
         &["check", "build.zen"][..],
         &["test", "build.zen"][..],
         &["emit", "build.zen"][..],
         &["build-graph", "build.zen"][..],
-    ] {
-        assert_build_zen_command_rejects_gated_target_field(args, field, value);
-    }
+    ]
 }
 
-fn assert_build_zen_command_rejects_gated_target_field(args: &[&str], field: &str, value: &str) {
+fn run_build_zen_command(
+    args: &[&str],
+    build_source: String,
+) -> (tempfile::TempDir, std::process::Output) {
     let tmp = tempfile::tempdir().expect("create temp dir");
-    std::fs::write(
-        tmp.path().join("build.zen"),
-        format!(
-            r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {{
-    b.add(Executable {{
-        name: "app",
-        main: "app.zen",
-        out_dir: "build/app/",
-        {field}: {value},
-    }})
-    .Ok(b.config())
-}}
-"#
-        ),
-    )
-    .expect("write build.zen");
+    std::fs::write(tmp.path().join("build.zen"), build_source).expect("write build.zen");
 
     let output = Command::new(env!("CARGO_BIN_EXE_zen"))
         .args(args)
@@ -157,6 +29,16 @@ build = (b: Builder) Result<BuildConfig, BuildError> {{
         .output()
         .expect("run zen build.zen command");
 
+    (tmp, output)
+}
+
+fn assert_rejected_without_outputs(
+    tmp: &tempfile::TempDir,
+    output: &std::process::Output,
+    args: &[&str],
+    expected: String,
+    reason: &str,
+) {
     assert!(
         !output.status.success(),
         "zen {args:?} unexpectedly succeeded: stdout={}, stderr={}",
@@ -164,72 +46,12 @@ build = (b: Builder) Result<BuildConfig, BuildError> {{
         String::from_utf8_lossy(&output.stderr)
     );
     assert!(
-        String::from_utf8_lossy(&output.stderr).contains(&format!(
-            "unsupported field `{field}` in `Executable` build target; package/link semantics are gated"
-        )),
-        "expected gated field diagnostic, stderr={}",
+        String::from_utf8_lossy(&output.stderr).contains(&expected),
+        "expected {reason} diagnostic, stderr={}",
         String::from_utf8_lossy(&output.stderr)
     );
     assert!(
         !tmp.path().join("build").exists(),
-        "zen {args:?} should reject gated target fields before creating build outputs"
-    );
-}
-
-fn assert_build_zen_commands_reject_unknown_target_field(field: &str, value: &str) {
-    for args in [
-        &["build", "build.zen"][..],
-        &["build.zen"][..],
-        &["check", "build.zen"][..],
-        &["test", "build.zen"][..],
-        &["emit", "build.zen"][..],
-        &["build-graph", "build.zen"][..],
-    ] {
-        assert_build_zen_command_rejects_unknown_target_field(args, field, value);
-    }
-}
-
-fn assert_build_zen_command_rejects_unknown_target_field(args: &[&str], field: &str, value: &str) {
-    let tmp = tempfile::tempdir().expect("create temp dir");
-    std::fs::write(
-        tmp.path().join("build.zen"),
-        format!(
-            r#"
-build = (b: Builder) Result<BuildConfig, BuildError> {{
-    b.add(Executable {{
-        name: "app",
-        main: "app.zen",
-        out_dir: "build/app/",
-        {field}: {value},
-    }})
-    .Ok(b.config())
-}}
-"#
-        ),
-    )
-    .expect("write build.zen");
-
-    let output = Command::new(env!("CARGO_BIN_EXE_zen"))
-        .args(args)
-        .current_dir(tmp.path())
-        .output()
-        .expect("run zen build.zen command");
-
-    assert!(
-        !output.status.success(),
-        "zen {args:?} unexpectedly succeeded: stdout={}, stderr={}",
-        String::from_utf8_lossy(&output.stdout),
-        String::from_utf8_lossy(&output.stderr)
-    );
-    assert!(
-        String::from_utf8_lossy(&output.stderr).contains(&format!(
-            "unknown field `{field}` in `Executable` build target"
-        )),
-        "expected unknown field diagnostic, stderr={}",
-        String::from_utf8_lossy(&output.stderr)
-    );
-    assert!(
-        !tmp.path().join("build").exists(),
-        "zen {args:?} should reject unknown target fields before creating build outputs"
+        "zen {args:?} should reject {reason} before creating build outputs"
     );
 }
